add a little dsl class now we don't need to create a separate utility methods for the test, instead we can write test right away
